Deprecate wsgi functions
The ``get_script_name``, ``get_query_string``, ``peek_path_info``,
``pop_path_info``, ``extract_path_info`` are not used anywhere in
Werkzeug and are likely not useful to others.
