+++ src/werkzeug/wsgi.py
@@ -1,6 +1,7 @@
 import io
 import re
 import typing as t
+import warnings
 from functools import partial
 from functools import update_wrapper
 from itertools import chain
@@ -11,6 +12,10 @@ from ._internal import _to_str
 from .sansio import utils as _sansio_utils
 from .sansio.utils import host_is_trusted  # noqa: F401 # Imported as part of API
 from .urls import _URLTuple
+from .urls import uri_to_iri
+from .urls import url_join
+from .urls import url_parse
+from .urls import url_quote
 
 if t.TYPE_CHECKING:
     from _typeshed.wsgi import WSGIApplication
@@ -173,9 +178,21 @@ def get_query_string(environ: "WSGIEnvironment") -> str:
 
     :param environ: WSGI environment to get the query string from.
 
+    .. deprecated:: 2.2
+        Will be removed in Werkzeug 2.3.
+
     .. versionadded:: 0.9
     """
-    return _sansio_utils.get_query_string(query_string=environ.get("QUERY_STRING", ""))
+    warnings.warn(
+        "'get_query_string' is deprecated and will be removed in Werkzeug 2.3.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    qs = environ.get("QUERY_STRING", "").encode("latin1")
+    # QUERY_STRING really should be ascii safe but some browsers
+    # will send us some unicode stuff (I am looking at you IE).
+    # In that case we want to urllib quote it badly.
+    return url_quote(qs, safe=":&%=+$!*'(),")
 
 
 def get_path_info(
@@ -206,8 +223,16 @@ def get_script_name(
         should be performed.
     :param errors: The decoding error handling.
 
+    .. deprecated:: 2.2
+        Will be removed in Werkzeug 2.3.
+
     .. versionadded:: 0.9
     """
+    warnings.warn(
+        "'get_script_name' is deprecated and will be removed in Werkzeug 2.3.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     path = environ.get("SCRIPT_NAME", "").encode("latin1")
     return _to_str(path, charset, errors, allow_none_charset=True)  # type: ignore
 
@@ -233,6 +258,9 @@ def pop_path_info(
     >>> env['SCRIPT_NAME']
     '/foo/a/b'
 
+    .. deprecated:: 2.2
+        Will be removed in Werkzeug 2.3.
+
     .. versionadded:: 0.5
 
     .. versionchanged:: 0.9
@@ -245,6 +273,12 @@ def pop_path_info(
     :param errors: The ``errors`` paramater passed to
         :func:`bytes.decode`.
     """
+    warnings.warn(
+        "'pop_path_info' is deprecated and will be removed in Werkzeug 2.3.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+
     path = environ.get("PATH_INFO")
     if not path:
         return None
@@ -285,6 +319,9 @@ def peek_path_info(
 
     If the `charset` is set to `None` bytes are returned.
 
+    .. deprecated:: 2.2
+        Will be removed in Werkzeug 2.3.
+
     .. versionadded:: 0.5
 
     .. versionchanged:: 0.9
@@ -293,6 +330,12 @@ def peek_path_info(
 
     :param environ: the WSGI environment that is checked.
     """
+    warnings.warn(
+        "'peek_path_info' is deprecated and will be removed in Werkzeug 2.3.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+
     segments = environ.get("PATH_INFO", "").lstrip("/").split("/", 1)
     if segments:
         return _to_str(  # type: ignore
@@ -340,23 +383,72 @@ def extract_path_info(
                                   same server point to the same
                                   resource.
 
+    .. deprecated:: 2.2
+        Will be removed in Werkzeug 2.3.
+
     .. versionchanged:: 0.15
         The ``errors`` parameter defaults to leaving invalid bytes
         quoted instead of replacing them.
 
     .. versionadded:: 0.6
+
     """
-    if isinstance(environ_or_baseurl, dict):
-        baseurl = get_current_url(environ_or_baseurl, root_only=True)
-    else:
-        baseurl = environ_or_baseurl
-    return _sansio_utils.extract_path_info(
-        baseurl=baseurl,
-        path_or_url=path_or_url,
-        charset=charset,
-        errors=errors,
-        collapse_http_schemes=collapse_http_schemes,
+    warnings.warn(
+        "'extract_path_info' is deprecated and will be removed in Werkzeug 2.3.",
+        DeprecationWarning,
+        stacklevel=2,
     )
+
+    def _normalize_netloc(scheme: str, netloc: str) -> str:
+        parts = netloc.split("@", 1)[-1].split(":", 1)
+        port: t.Optional[str]
+
+        if len(parts) == 2:
+            netloc, port = parts
+            if (scheme == "http" and port == "80") or (
+                scheme == "https" and port == "443"
+            ):
+                port = None
+        else:
+            netloc = parts[0]
+            port = None
+
+        if port is not None:
+            netloc += f":{port}"
+
+        return netloc
+
+    # make sure whatever we are working on is a IRI and parse it
+    path = uri_to_iri(path_or_url, charset, errors)
+    if isinstance(environ_or_baseurl, dict):
+        environ_or_baseurl = get_current_url(environ_or_baseurl, root_only=True)
+    base_iri = uri_to_iri(environ_or_baseurl, charset, errors)
+    base_scheme, base_netloc, base_path = url_parse(base_iri)[:3]
+    cur_scheme, cur_netloc, cur_path = url_parse(url_join(base_iri, path))[:3]
+
+    # normalize the network location
+    base_netloc = _normalize_netloc(base_scheme, base_netloc)
+    cur_netloc = _normalize_netloc(cur_scheme, cur_netloc)
+
+    # is that IRI even on a known HTTP scheme?
+    if collapse_http_schemes:
+        for scheme in base_scheme, cur_scheme:
+            if scheme not in ("http", "https"):
+                return None
+    else:
+        if not (base_scheme in ("http", "https") and base_scheme == cur_scheme):
+            return None
+
+    # are the netlocs compatible?
+    if base_netloc != cur_netloc:
+        return None
+
+    # are we below the application path?
+    base_path = base_path.rstrip("/")
+    if not cur_path.startswith(base_path):
+        return None
+
+    return f"/{cur_path[len(base_path) :].lstrip('/')}"
 
 
 class ClosingIterator:
